lisa.tests.base: Return float in RTATestBundle.unscaled_utilization()
Instead of rounding down the percentage, return a float. lisa.wlgen.rta will
round the values at the step to avoid precision loss as much as possible.

+++ lisa/tests/base.py
@@ -1334,7 +1334,7 @@ class RTATestBundle(FtraceTestBundle, DmesgTestBundle):
         else:
             capacity_scale = 1024
 
-        return int((plat_info["cpu-capacities"]['rtapp'][cpu] / capacity_scale) * utilization_pct)
+        return (plat_info["cpu-capacities"]['rtapp'][cpu] / capacity_scale) * utilization_pct
 
     @classmethod
     @abc.abstractmethod
